requirements.txt
Django==1.6.2
Fabric==1.8.2
MySQL-python==1.2.3
Pillow==2.3.0
Pygments==1.6
South==0.8.4
Werkzeug==0.9.4
argparse==1.2.1
astroid==1.0.1
clonedigger==1.1.0
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.5
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-jenkins==0.15.0
django-pipeline==1.3.22
django-toolbelt==0.0.1
djangorestframework==2.3.12
ecdsa==0.10
epydoc==3.0.1
futures==2.1.6
gunicorn==18.0
jslint==0.6.0
logilab-common==0.61.0
lxml==3.3.3
paramiko==1.12.2
pep8==1.4.6
psycopg2==2.5.2
pycallgraph==1.0.1
pycrypto==2.6.1
pyflakes==0.7.3
pylint==1.1.0
pymongo==2.6.3
pystache==0.5.3
selenium==2.40.0
six==1.5.2
sqlparse==0.1.11
static==1.0.2
wsgiref==0.1.2
